+++ charts/platform/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this chart will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.27.3] - 2026-03-03
+
+### Changed
+
+- Updated platform chart version to 0.27.3
+- Updated platform app version to v25.3.3
+- Updated studios subchart to version 1.2.1
+- Updated pipeline-optimization subchart to version 1.0.1
+
 ## [0.27.0] - 2026-02-20
 
 ### Added