+++ charts/platform/charts/pipeline-optimization/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this chart will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2026-03-03
+
+### Changed
+
+- Updated chart version to 1.0.1 to mark stable release of pipeline-optimization subchart and update
+  app version to latest patch version 0.4.11
+
 ## [0.3.0] - 2026-02-20
 
 ### Added